requirements.txt
aiogram
python-dotenv
openai
redis